closer to typechecking
seems like we’ll need some kind of Variant type to handle the attributes (but maybe this doesn’t have to be too hard? question is how much it uglies up the construction of programs.
If each model has to have a lot of defined accessors, that might not be the worst actually…